// apps/serverless-admin/src/lib/schemas.ts
export const codeArtifactSchema = z.object({
  bucket: z.string().optional(),
  key: z.string(),
  sha256: z.string().optional(),
  sizeBytes: z.number().optional(),
  source: z.string().optional(),
})
// ...
export const layerVersionSchema = z.object({
  id: z.string(),
  name: z.string(),
  version: z.number(),
  layerArn: z.string().optional(),
  description: z.string().optional(),
  codeArtifact: codeArtifactSchema.nullish(),
  compatibleRuntimes: z.array(z.string()).optional(),
  compatibleArchitectures: z.array(z.string()).optional(),
  createdAt: z.string().optional(),
})
// ...
export type LayerVersion = z.infer<typeof layerVersionSchema>